+++ src/main/org/firebirdsql/gds/ng/FbStatement.java
@@ -69,8 +69,15 @@ public interface FbStatement extends ExceptionListenable {
 
     /**
      * @return descriptor of the fields returned by this statement
+     * @deprecated Use {@link #getRowDescriptor()} instead; will be removed in Jaybird 5
      */
+    @Deprecated
     RowDescriptor getFieldDescriptor();
+
+    /**
+     * @return descriptor of the row returned by this statement
+     */
+    RowDescriptor getRowDescriptor();
 
     /**
      * @return The statement type